Cross-browser workaround for CSS pointer-events property on hero image
When the hero image is hovered, the script checks whether the mouse is
within the bounding box of the third membership option. If it is, a
class is set to simulate hovering. The script also makes sure that the
box is clickable even where it is overlapped by the hero image.